+++ transportation-service/src/main/java/com/ftcs/transportation/schelude/service/ScheduleService.java
@@ -44,7 +44,7 @@ public class ScheduleService {
 
 
 
-    public void updateSchedule(ScheduleRequestDTO requestDTO, Integer scheduleId) {
+    public void updateSchedule(ScheduleRequestDTO requestDTO, Long scheduleId) {
         Schedule schedule = getScheduleById(scheduleId);
         validateScheduleDates(requestDTO.getStartDate(), requestDTO.getEndDate());
         mapScheduleRequestToEntity(requestDTO, schedule);
@@ -52,7 +52,7 @@ public class ScheduleService {
         scheduleRepository.save(schedule);
     }
 
-    public void updateStatusSchedule(UpdateStatusScheduleRequestDTO requestDTO, Integer scheduleId) {
+    public void updateStatusSchedule(UpdateStatusScheduleRequestDTO requestDTO, Long scheduleId) {
         Schedule schedule = getScheduleById(scheduleId);
         schedule.setStatus(requestDTO.getStatus());
         schedule.setUpdateAt(LocalDateTime.now());
@@ -75,7 +75,7 @@ public class ScheduleService {
         return schedules;
     }
 
-    public Schedule getScheduleById(Integer scheduleId) {
+    public Schedule getScheduleById(Long scheduleId) {
         return scheduleRepository.findScheduleByScheduleId(scheduleId)
                 .orElseThrow(() -> new BadRequestException("Schedule does not exist"));
     }
